Struct coredb::index_manager::index::Index

pub struct Index {
    metadata: Metadata,
    all_segments_map: DashMap<u32, Segment>,
    index_dir_path: String,
    index_dir_lock: Arc<Mutex<ThreadId>>,
}

Index for storing log messages and metric points.

Fields
metadata: Metadata

Metadata for this index.

all_segments_map: DashMap<u32, Segment>

DashMap of segment number to segment.

index_dir_path: String

Directory where the index is serialized.

index_dir_lock: Arc<Mutex<ThreadId>>

Mutex for locking the directory where the index is committed / read from, so that two threads don’t write the directory at the same time.

Implementations

impl Index

pub fn new(index_dir_path: &str) -> Result<Self, CoreDBError>

Create a new index with default threshold log messages / threshold metric points parameters. However, if a directory with the same path already exists and has a metadata file in it, the function will refresh the existing index instead of creating a new one. If the refresh process fails, an error will be thrown to indicate the issue.

pub fn new_with_threshold_params( index_dir: &str, num_log_messages_threshold: u32, num_metric_points_threshold: u32 ) -> Result<Self, CoreDBError>

Creates a new index at a specified directory path with customizable parameters for log message and metric point thresholds. If a directory with the same path already exists and has a metadata file in it, the existing index will be refreshed instead of creating a new one. If the refresh process fails, an error will be thrown to indicate the issue.

fn get_current_segment_ref(&self) -> Ref<'_, u32, Segment>

Get the reference for the current segment.

pub fn append_log_message( &self, time: u64, fields: &HashMap<String, String>, message: &str )

Append a log message to the current segment of the index.

pub fn append_metric_point( &self, metric_name: &str, labels: &HashMap<String, String>, time: u64, value: f64 )

Append a metric point to the current segment of the index.

pub fn search_logs( &self, query: &str, range_start_time: u64, range_end_time: u64 ) -> Vec<LogMessage>

Search for given query in the given time range.

fn commit_segment(&self, segment_number: u32, sync_after_write: bool)

Helper function to commit a segment with given segment_number to disk.

fn is_too_big(&self, segment_number: u32) -> bool

Helper function to check if the given segment is ‘too big’. In other words, it returns ‘true’ if the number of log messages or the number of metric points in this segment exceeded the corresponding max values.

pub fn commit(&self, sync_after_write: bool)

Commit a segment to disk.

If sync_after_write is set to true, make sure that the OS buffers are flushed to disk before returning (typically sync_after_write should be set to true in tests that refresh the index immediately after committing).

pub fn refresh(index_dir_path: &str) -> Result<Self, CoreDBError>

Read the index from the given index_dir_path.

pub fn get_overlapping_segments( &self, range_start_time: u64, range_end_time: u64 ) -> Vec<u32>

Returns segment numbers of segments that overlap with the given time range.

pub fn get_metrics( &self, label_name: &str, label_value: &str, range_start_time: u64, range_end_time: u64 ) -> Vec<MetricPoint>

Get metric points corresponding to given label name and value, within the given range (inclusive of both start and end time).

pub fn get_index_dir(&self) -> String

pub fn delete(&self)

Function to delete the index directory.